# apps/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MyHR API",
...
    "version": "0.0.1"
  },
  "components": {
    "securitySchemes": {
      "masterApiKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "Master API key issued to 1tap (env `MASTER_API_KEY`). Sent as `Authorization: Bearer mh_live_…`. Master callers may operate across all tenants and must send `X-Tenant-Id` to scope tenant-specific calls."
      },
      "tenantApiKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "Org-scoped API key minted from the dashboard. Sent as `Authorization: Bearer mh_live_…`. The org id is derived from the key, so `X-Tenant-Id` is ignored."
      },
      "userSession": {
        "type": "http",
        "scheme": "bearer",
        "description": "Better Auth session token forwarded by the web app. End-user callers send `X-Org-Id` to select which of their orgs the request targets. `X-Actor` is ignored on user-session requests (the actor is taken from the session to prevent spoofing)."
      }
    },
...
  },
  "paths": {
...
    "/v1/invitations/accept": {
      "post": {
        "operationId": "acceptInvitation",
        "summary": "Accept an invitation",
        "tags": [
          "Invitations"
        ],
        "description": "Exchanges an invitation token for a membership. The caller's email must match the email on the invitation.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string",
                    "minLength": 1
                  }
                },
                "required": [
                  "token"
                ],
                "additionalProperties": false
              },
              "example": {
                "token": "inv_4f3c1aa9e2b14d8e9c0f7d6b5e2a1c8d"
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 200
            },
            "in": "header",
            "name": "idempotency-key",
            "required": true,
            "description": "Required on all writes (POST/PATCH/DELETE). Replays return the cached response. Reusing a key with a different body returns 409."
          }
        ],
        "security": [
          {
            "userSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "orgId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "userId": {
                      "type": "string"
                    },
                    "role": {
                      "type": "string",
                      "enum": [
                        "owner",
                        "admin",
                        "manager",
                        "member"
                      ]
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": [
                    "id",
                    "orgId",
                    "userId",
                    "role",
                    "createdAt"
                  ],
                  "additionalProperties": false
                },
                "example": {
                  "id": "cccccccc-dddd-eeee-ffff-aaaaaaaaaaaa",
                  "orgId": "11111111-2222-3333-4444-555555555555",
                  "userId": "u_01HXYZUSER00000000000000",
                  "role": "admin",
                  "createdAt": "2026-05-04T12:00:00.000Z"
                }
              }
            }
          },
          "400": {
            "description": "Validation or request shape error.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "description": "Machine-readable error code, e.g. `not_found`."
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {}
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false,
                  "description": "Validation or request shape error."
                },
                "example": {
                  "error": {
                    "code": "bad_request",
                    "message": "managerId does not reference an employee in this tenant"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "description": "Machine-readable error code, e.g. `not_found`."
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {}
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false,
                  "description": "Missing or invalid API key."
                },
                "example": {
                  "error": {
                    "code": "unauthorized",
                    "message": "Missing or invalid API key"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Caller is authenticated but not allowed to perform this action.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "description": "Machine-readable error code, e.g. `not_found`."
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {}
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false,
                  "description": "Caller is authenticated but not allowed to perform this action."
                },
                "example": {
                  "error": {
                    "code": "forbidden",
                    "message": "Caller is not allowed to perform this action"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Resource not found.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "description": "Machine-readable error code, e.g. `not_found`."
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {}
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false,
                  "description": "Resource not found."
                },
                "example": {
                  "error": {
                    "code": "not_found",
                    "message": "Not found"
                  }
                }
              }
            }
          },
          "409": {
            "description": "Conflict — e.g. unique constraint violation, or Idempotency-Key reused with a different body.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "description": "Machine-readable error code, e.g. `not_found`."
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {}
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false,
                  "description": "Conflict — e.g. unique constraint violation, or Idempotency-Key reused with a different body."
                },
                "example": {
                  "error": {
                    "code": "conflict",
                    "message": "An employee with that email or external_id already exists"
                  }
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded. Retry after the seconds indicated by the `Retry-After` header.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "description": "Machine-readable error code, e.g. `not_found`."
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {}
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false,
                  "description": "Rate limit exceeded. Retry after the seconds indicated by the `Retry-After` header."
                },
                "example": {
                  "error": {
                    "code": "rate_limited",
                    "message": "Rate limit exceeded. Retry after 6s.",
                    "details": {
                      "retryAfterSec": 6
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "description": "Machine-readable error code, e.g. `not_found`."
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {}
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false,
                  "description": "Unexpected server error."
                },
                "example": {
                  "error": {
                    "code": "internal_error",
                    "message": "Internal server error"
                  }
                }
              }
            }
          }
        },
        "x-codeSamples": [
          {
            "lang": "bash",
            "label": "curl",
            "source": "curl -X POST 'https://api.myhr.example/v1/invitations/accept' \\\n  -H 'Authorization: Bearer $MYHR_API_KEY' \\\n  -H 'Idempotency-Key: '\"$(uuidgen)\" \\\n  -H 'Content-Type: application/json' \\\n  -d '{\"token\":\"inv_4f3c1aa9e2b14d8e9c0f7d6b5e2a1c8d\"}'"
          },
          {
            "lang": "javascript",
            "label": "@myhr/sdk",
            "source": "import { createClient } from \"@myhr/sdk\";\n\nconst myhr = createClient({\n  baseUrl: \"https://api.myhr.example\",\n  getToken: () => process.env.MYHR_API_KEY!,\n  defaultTenantId: \"11111111-2222-3333-4444-555555555555\",\n});\n\nawait myhr.invitations.accept(\"inv_4f3c1aa9e2b14d8e9c0f7d6b5e2a1c8d\");"
          }
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "https://api.myhr.example",
      "description": "test"
    }
  ],
  "security": [
    {
      "masterApiKey": []
    },
    {
      "tenantApiKey": []
    },
    {
      "userSession": []
    }
  ],
...
}